+++ Basil/include/Scene/Scene.h
@@ -25,6 +25,7 @@ namespace Basil
 			uint32_t viewportWidth;
 			uint32_t viewportHeight;
 			friend class Entity;
+			friend class PropertiesPanel;
 			friend class SceneHierarchyPanel;
 	};
 }